refactor: Move `getAllTasksByOwnerId`, `getAllTasksByAssigneeId` to `TaskController`
